requirements.txt
# ----------------------------------------------------------------------------
# 环境变量管理
# ----------------------------------------------------------------------------

# 环境变量加载
python-dotenv>=1.0.0

# ----------------------------------------------------------------------------
# Langchain 核心包（1.0 版本）
# ----------------------------------------------------------------------------

# Langchain 核心库
langchain>=1.0.0,<2.0.0
langchain-core>=1.0.0,<2.0.0

